Added missing cast for llvm-gcc.

+++ src/appleseed/renderer/modeling/environmentedf/hosekenvironmentedf.cpp
@@ -213,7 +213,7 @@ namespace
         {
             const double clamped_turbidity = clamp(turbidity, 1.0, 10.0) - 1.0;
             const size_t turbidity_low = truncate<size_t>(clamped_turbidity);
-            const size_t turbidity_high = min(turbidity_low + 1, 9u);
+            const size_t turbidity_high = min(turbidity_low + 1, size_t(9));
             const double turbidity_interp = clamped_turbidity - turbidity_low;
 
             for (size_t w = 0; w < 3; ++w)
